+++ src/ResourceManager/GuestConfiguration/Commands.GuestConfiguration.Test/ScenarioTests/GetVmGuestPolicyStatusHistoryTests.cs
@@ -20,11 +20,11 @@ namespace Microsoft.Azure.Commands.GuestConfiguration.Test.ScenarioTests
     using Xunit;
 
 
-    public class GetVmGuestPolicyReportTests
+    public class GetVmGuestPolicyStatusHistoryTests
     {
         private readonly XunitTracingInterceptor _logger;
 
-        public GetVmGuestPolicyReportTests(Xunit.Abstractions.ITestOutputHelper output)
+        public GetVmGuestPolicyStatusHistoryTests(Xunit.Abstractions.ITestOutputHelper output)
         {
             _logger = new XunitTracingInterceptor(output);
             XunitTracingInterceptor.AddToContext(_logger);
@@ -35,28 +35,28 @@ namespace Microsoft.Azure.Commands.GuestConfiguration.Test.ScenarioTests
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void VmNameScope()
         {
-            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyReport-VmNameScope");
+            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyStatusHistory-VmNameScope");
         }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void InitiativeIdScope()
         {
-            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyReport-InitiativeIdScope");
+            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyStatusHistory-InitiativeIdScope");
         }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void InitiativeNameScope()
         {
-            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyReport-InitiativeNameScope");
+            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyStatusHistory-InitiativeNameScope");
         }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void ReportIdScope()
+        public void ShowOnlyChangeSwitchVmNameScope()
         {
-            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyReport-ReportIdScope");
+            TestController.NewInstance.RunPowerShellTest(_logger, "Get-AzureRmVMGuestPolicyStatusHistory-ShowOnlyChangeSwitch-VmNameScope");
         }
     }
 }